Radom apples is nor working

+++ TaskUtils.h
@@ -8,7 +8,7 @@ public:
 	static const char* _bc_path;
 	static const char* _apple_path;
 
-	static int RandomInt(size_t range = 0);
+	static int RandomInt(int range = 0);
 	static char KeyPressed();
 };
